requirements.txt
openai
langchain
faiss-cpu
PyPDF2
python-docx
pandas
numpy
python-dotenv
